requirements.txt
yt-dlp
openai
python-dotenv
tqdm